Added an alternate (commented out) allergy function.
I made these changes a couple of days ago and forgot to actually push them. What this version of the allergy function does is find the allergies from the given pizza (hardcoded as a sausage pizza for now) and display those on the allergies page. While it is functional, I'm not sure how compatible it is with the current changes, or if it's even useful at all, so I left it commented out for the time being.